SOA_hw_2/user_service/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
SQLAlchemy==1.4.46
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
pytest==7.2.1
PyJWT==2.6.0
pydantic==1.10.7
email-validator==1.3.1